Extract shared methods/scopes into a BaseExecution to share between Execution and Job models; remove deprecated ActiveJobJob model
