+++ app/filters/good_job/base_filter.rb
@@ -32,7 +32,7 @@ module GoodJob
 
     def job_classes
       filtered_query(params.slice(:queue_name)).unscope(:select)
-                                               .group("serialized_params->>'job_class'").count
+                                               .group(GoodJob::BaseExecution.params_job_class).count
                                                .sort_by { |name, _count| name.to_s }
                                                .to_h
     end
