--- spec/app/models/good_job/active_job_job_spec.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-require 'rails_helper'
-
-RSpec.describe GoodJob::ActiveJobJob do
-  describe '#initialize' do
-    it 'is deprecated' do
-      allow(ActiveSupport::Deprecation).to receive(:warn)
-      described_class.create!
-      expect(ActiveSupport::Deprecation).to have_received(:warn)
-    end
-  end
-end
